api/requirements.txt
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
cftime==1.6.2
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.6
click-plugins==1.1.1
cligj==0.7.2
dotty-dict==1.3.1
Fiona==1.9.4.post1
Flask==2.3.2
Flask-Cors==4.0.0
geopandas==0.13.2
gunicorn==21.1.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
mailchimp-marketing==3.0.80
MarkupSafe==2.1.3
netCDF4==1.6.4
numpy==1.25.1
packaging==23.1
pandas==2.0.3
pathspec==0.11.1
pyproj==3.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
shapely==2.0.1
six==1.16.0
tomli==2.0.1
tzdata==2023.3
urllib3==2.0.3
Werkzeug==2.3.6
zipp==3.16.2
